requirements.txt
robobrowser==0.5.3
pytesseract==0.2.2
lxml==4.2.1
httplib2==0.11.3
nltk==3.3
Pillow==5.2.0
beautifulsoup4==4.6.3
